+++ app/infrastructure/services/dependencies.py
@@ -8,19 +8,15 @@ from typing import Annotated
 
 from fastapi import Depends
 
-from infrastructure.clients.google_workspace import GoogleWorkspaceClients
 from infrastructure.clients.maxmind import MaxMindClient
 from infrastructure.configuration import AppSettings, Settings
-from infrastructure.directory.provider import DirectoryProvider
 from infrastructure.events.service import EventDispatcher
 from infrastructure.i18n.service import TranslationService
 from infrastructure.idempotency.protocol import IdempotencyService
 from infrastructure.resilience.service import ResilienceService
 from infrastructure.services.providers import (
     get_app_settings,
-    get_directory_provider,
     get_event_dispatcher,
-    get_google_workspace_clients,
     get_idempotency_service,
     get_maxmind_client,
     get_resilience_service,
@@ -32,12 +28,6 @@ from infrastructure.services.providers import (
 SettingsDep = Annotated[Settings, Depends(get_settings)]
 AppSettingsDep = Annotated[AppSettings, Depends(get_app_settings)]
 
-# Google Workspace clients facade dependency - provides attribute-based access to all Google services
-# Usage: google.directory.list_groups(), google.drive.create_file(...), etc.
-# For types and data classes: from infrastructure.clients.google_workspace import ...
-GoogleWorkspaceClientsDep = Annotated[
-    GoogleWorkspaceClients, Depends(get_google_workspace_clients)
-]
 
 # MaxMind client dependency - provides IP geolocation operations
 # Usage: maxmind.geolocate(ip_address="8.8.8.8")
@@ -56,16 +46,11 @@ IdempotencyServiceDep = Annotated[IdempotencyService, Depends(get_idempotency_se
 # Resilience service dependency
 ResilienceServiceDep = Annotated[ResilienceService, Depends(get_resilience_service)]
 
-# Directory provider dependency — IDP-agnostic group membership and listing
-DirectoryProviderDep = Annotated[DirectoryProvider, Depends(get_directory_provider)]
-
 __all__ = [
     "SettingsDep",
     "AppSettingsDep",
-    "GoogleWorkspaceClientsDep",
     "EventDispatcherDep",
     "TranslationServiceDep",
     "IdempotencyServiceDep",
     "ResilienceServiceDep",
-    "DirectoryProviderDep",
 ]
