+++ app/packages/access/request/providers.py
@@ -10,9 +10,9 @@ module scope — never bypass providers by instantiating services directly.
 from functools import lru_cache
 
 from infrastructure.services import (
-    get_directory_provider,
     get_event_dispatcher,
 )
+from infrastructure.directory import get_directory_provider
 from infrastructure.storage import get_storage_service
 from packages.access.request.service import AccessRequestService
 from packages.access.request.store import AccessRequestRepository
