+++ app/packages/access/sync/providers.py
@@ -3,9 +3,9 @@ from typing import Dict
 
 from infrastructure.clients.aws import get_aws_clients
 from infrastructure.services import (
-    get_directory_provider,
     get_event_dispatcher,
 )
+from infrastructure.directory import get_directory_provider
 from infrastructure.storage import get_storage_service
 from packages.access.common.providers import get_access_runtime_config
 from packages.access.common.settings import AccessSyncSettings, get_access_settings
