+++ app/server/lifespan.py
@@ -8,6 +8,11 @@ from fastapi import FastAPI
 from slack_bolt import App
 from structlog.stdlib import BoundLogger
 
+from infrastructure.configuration.infrastructure.server import (
+    ServerSettings,
+    get_server_settings,
+)
+from infrastructure.directory import get_directory_provider
 from infrastructure.i18n.resources import I18nResourceSpec
 from infrastructure.logging.setup import configure_logging
 from infrastructure.platforms import get_platform_service
@@ -16,12 +21,7 @@ from infrastructure.plugins import (
     register_feature_integrations,
 )
 from infrastructure.security import get_jwks_manager
-from infrastructure.configuration.infrastructure.server import (
-    get_server_settings,
-    ServerSettings,
-)
 from infrastructure.services import (
-    get_directory_provider,
     get_settings,
     get_translation_service,
 )
